double revision

+++ examples/H2/train.py
@@ -29,7 +29,7 @@ args = parser.parse_args()
 name = args.name
 
 # Do everything on the CPU in double precision.
-dtype = pt.float32
+dtype = pt.float64
 device = pt.device( device_str )
 gen = pt.Generator( device=device )
